Also update testcase/example/simple for long `yychar`.
And remove old `Yyxlex.java` which was submitted (probably by mistake) in commit 4db21cb8fbf6d06d6cffee8a4a086e901d60f604

+++ testsuite/testcases/src/test/cases/simple/Yytoken.java
@@ -1,12 +1,13 @@
 public class Yytoken {
+
   public int m_index;
   public String m_text;
   public int m_line;
-  public int m_charBegin;
-  public int m_charEnd;
-  
-  Yytoken (int index, String text, int line, int charBegin, int charEnd) {
-     m_index = index;
+  public long m_charBegin;
+  public long m_charEnd;
+
+  Yytoken(int index, String text, int line, long charBegin, long charEnd) {
+    m_index = index;
     m_text = text;
     m_line = line;
     m_charBegin = charBegin;
@@ -14,11 +15,15 @@ public class Yytoken {
   }
 
   public String toString() {
-    return "Text   : "+m_text+
-           "\nindex : "+m_index+
-           "\nline  : "+m_line+
-           "\ncBeg. : "+m_charBegin+
-           "\ncEnd. : "+m_charEnd;
+    return "Text   : "
+        + m_text
+        + "\nindex : "
+        + m_index
+        + "\nline  : "
+        + m_line
+        + "\ncBeg. : "
+        + m_charBegin
+        + "\ncEnd. : "
+        + m_charEnd;
   }
 }
-
